Rename 'person' package to 'client' package in test/
